backend/requirements.txt
# ScholarAI Backend Dependencies

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Authentication
Flask-JWT-Extended==4.6.0

# Database
pymongo==4.6.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
feedparser==6.0.10

# Security
cryptography==41.0.7

# Testing (optional, for development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
